+++ src/index.ts
@@ -1,5 +1,7 @@
-import GMStorage from './gm-storage.js'
+import { GMStorage } from './gm-storage.js'
 
 export default GMStorage
-export type { Callback, Options } from './gm-storage.js'
+export { GMStorage, GMStorage as GMStore }
+export { JSONKeyStorage, JSONKeyStorage as JSONKeyStore } from './json-key-storage.js'
+export type { Callback, Options } from './gm-storage-base.js'
 export type { JSONValue } from './util.js'